chore: refactoring mutations, queries and types to APIClients folder (#98)
* refactoring all mutations and queries to APIClients folder

* refactor request and response types to apiclients folder
