+++ frontend/src/APIClients/types/UserClientTypes.ts
@@ -1,0 +1,17 @@
+import { Role } from "../../types/AuthTypes";
+
+export type UserRequest = {
+  firstName: string;
+  lastName: string;
+  password: string;
+  email: string;
+  role: Role;
+};
+
+export type UserResponse = {
+  id: string;
+  firstName: string;
+  lastName: string;
+  email: string;
+  role: Role;
+};
