+++ frontend/src/components/user-management/AddAdminModal.tsx
@@ -27,8 +27,8 @@ import AdminConfirmationMessage from "./AdminConfirmationMessage";
 import ModalFooterButtons from "../common/ModalFooterButtons";
 import { PlusOutlineIcon } from "../../assets/icons";
 import { ADMIN_PAGE } from "../../constants/Routes";
-import { UserRequest } from "../../types/UserTypes";
-import ADD_USER from "../../APIClients/mutations/UserMutations";
+import { UserRequest } from "../../APIClients/types/UserClientTypes";
+import { ADD_USER } from "../../APIClients/mutations/UserMutations";
 import GET_USERS_BY_ROLE from "../../APIClients/queries/UserQueries";
 
 const AddAdminModal = (): React.ReactElement => {
